--- a/FEATURE_SUMMARY.md
+++ b/FEATURE_SUMMARY.md
@@ -1,0 +1,1 @@
+# Ultra-Fast Voice Pipeline